+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [5.0.0] - 2021-02-22
+## [5.0.0] - 2021-02-24
 ### Added
- - Added Wordpress Code Standard
-
+ - Compatibility with WooCommerce v5.0.0
+ - Compatibility with WordPress v5.6.2
+ - Added Wordpress Code Standard at plugin
+### Changed
+ - Fixed round amount
 ## [4.6.4] - 2021-02-11
 ### Changed
  - Removed payments methods in option custom checkout OFF